migrating replacer configuration
