+++ src/domain/plainjs-template/template/index.js
@@ -6,12 +6,17 @@ import {
     printBody
 } from 'atherdon-old-newsletter-js-outertemplate';
 
-const plainV2 = () => {
-// 
 
-    console.log(printMain());
-    console.log(printFooter());
-    console.log(printBody());
+import TObject from 'atherdon-old-newsletter-js-outertemplate';
+
+const plainV2 = () => {
+
+
+
+
+    console.log(TObject.printMain());
+    console.log(TObject.printFooter());
+    console.log(TObject.printBody());
 
 
 }
