+++ src/domain/replacer-class/configuration-nmtg.js
@@ -1,0 +1,1 @@
+//--------------------
