+++ src/domain/replacer-class/configuration-react.js
@@ -1,0 +1,1 @@
+//--------------------
